all math functions working
