+++ javascripts/components/calculator.js
@@ -1,11 +1,6 @@
-const printToDom = (stringToPrint, divId) => {
-    const selectedDiv = document.getElementById(divId);
-    selectedDiv.innerHTML += stringToPrint;
-};
+import {printToDom} from "../helpers/util.js";
+import {multiply, divide, subtract, add} from "../helpers/maths.js";
 
-const multiply = (n1, n2) => {
-    return n1 * n2;
-}
 
 const calculate = (num1, num2, mathType) => {
     let answer = 0;
@@ -14,18 +9,18 @@ const calculate = (num1, num2, mathType) => {
             answer = multiply(num1, num2);
             break;
         case "divide":
-            answer = num1 / num2
+            answer = divide(num1, num2);
             break;
         case "add":
-            answer = num1 + num2
+            answer = add(num1, num2);
             break; 
         case "subtract":
-            answer = num1 - num2;
+            answer = subtract(num1, num2);
             break;
         default:
         answer = "nope";
     }
     printToDom(answer, "result");
-}
+};
 
-calculate(4, 2, "multiply");
+export {calculate};
