+++ javascripts/helpers/maths.js
@@ -1,0 +1,20 @@
+const multiply = (n1, n2) => {
+    return n1 * n2;
+};
+
+const divide = (n1, n2) => {
+    return n1 / n2;
+};
+
+const subtract = (n1, n2) => {
+    return n1 - n2;
+};
+
+const add = (n1, n2) => {
+    return n1 + n2;
+};
+
+export{multiply}
+export{divide}
+export{subtract}
+export{add}
