+++ javascripts/helpers/util.js
@@ -1,0 +1,8 @@
+const printToDom = (stringToPrint, divId) => {
+    const selectedDiv = document.getElementById(divId);
+    selectedDiv.innerHTML += stringToPrint;
+};
+
+
+// export always goes on the bottom
+export{printToDom}
