+++ javascripts/main.js
@@ -1,0 +1,6 @@
+import {calculate} from "./components/calculator.js";
+
+// calculate(4, 2, "multiply");
+// calculate(4, 2, "divide");
+// calculate(4, 2, "subtract");
+calculate(4, 2, "add");
